requirements.txt
alpaca-trade-api>=2.3.0
yfinance>=0.2.28
pandas>=1.5.3
numpy>=1.23.5
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
transformers==4.36.2
plotly==5.18.0 